Float/Double support for scale

+++ ImageToText/C++/ImageToText.h
@@ -2,4 +2,4 @@
 #include <fstream>
 #include <string>
 
-int imageToText(std::string& filename, unsigned int* scale, std::ofstream& output);
+int imageToText(std::string& filename, double* scale, std::ofstream& output);
